+++ apps/ledger-live-mobile/src/components/RootNavigator/BuyDeviceNavigator.tsx
@@ -2,10 +2,10 @@ import React, { useMemo } from "react";
 import { createStackNavigator } from "@react-navigation/stack";
 import { useTheme } from "styled-components/native";
 import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
-
 import { ScreenName } from "~/const";
 import { getStackNavigatorConfig } from "~/navigation/navigatorConfig";
 import GetDevice from "~/screens/GetDeviceScreen";
+import GetFlex from "LLM/features/Reborn/screens/UpsellFlex";
 import PurchaseDevice from "~/screens/PurchaseDevice";
 import { BuyDeviceNavigatorParamList } from "./types/BuyDeviceNavigator";
 
@@ -14,11 +14,15 @@ const Stack = createStackNavigator<BuyDeviceNavigatorParamList>();
 const BuyDeviceNavigator = () => {
   const { colors } = useTheme();
   const buyDeviceFromLive = useFeature("buyDeviceFromLive");
+  const upsellFlexFF = useFeature("llmRebornFlex");
   const stackNavigationConfig = useMemo(() => getStackNavigatorConfig(colors, true), [colors]);
 
   return (
     <Stack.Navigator screenOptions={{ ...stackNavigationConfig, headerShown: false }}>
-      <Stack.Screen name={ScreenName.GetDevice} component={GetDevice} />
+      <Stack.Screen
+        name={ScreenName.GetDevice}
+        component={upsellFlexFF?.enabled ? GetFlex : GetDevice}
+      />
       {buyDeviceFromLive?.enabled && (
         <Stack.Screen name={ScreenName.PurchaseDevice} component={PurchaseDevice} />
       )}