+++ libs/coin-modules/coin-module-boilerplate/src/index.ts
@@ -1,0 +1,4 @@
+export * from "./types";
+
+export { createBridges } from "./bridge/index";
+export type { BoilerplateCoinConfig } from "./config";